requirements.txt
streamlit>=1.32.0
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.12.0
ta>=0.11.0
